+++ core/configure/proto/multi_lang_general_model_service.proto
@@ -49,8 +49,8 @@ message EmptyRequest{
 };
 
 message ServingConfig{
-    int32 max_batch_size = 1;
-    string proto_txt = 2;
+    required int32 max_batch_size = 1;
+    required string proto_txt = 2;
 };
 
 service MultiLangGeneralModelService {
