+++ python/examples/fit_a_line/get_data.sh
