+++ python/examples/fit_a_line/test_multilang_client.py
@@ -17,8 +17,8 @@ from paddle_serving_client import MultiLangClient
 import sys
 
 client = MultiLangClient()
-client.load_client_config(sys.argv[1])
-client.connect(["127.0.0.1:9393"])
+#client.load_client_config(sys.argv[1])
+client.connect(["127.0.0.1:19293"])
 
 import paddle
 test_reader = paddle.batch(
