+++ python/examples/fit_a_line/test_multilang_server.py
@@ -15,9 +15,9 @@
 
 import os
 import sys
-from paddle_serving_server import OpMaker
-from paddle_serving_server import OpSeqMaker
-from paddle_serving_server import MultiLangServer
+from paddle_serving_server_gpu import OpMaker
+from paddle_serving_server_gpu import OpSeqMaker
+from paddle_serving_server_gpu import MultiLangServer
 
 op_maker = OpMaker()
 read_op = op_maker.create('general_reader')
@@ -32,5 +32,5 @@ op_seq_maker.add_op(response_op)
 server = MultiLangServer()
 server.set_op_sequence(op_seq_maker.get_op_sequence())
 server.load_model_config(sys.argv[1])
-server.prepare_server(workdir="work_dir1", port=9393, device="cpu")
+server.prepare_server(workdir="test19293", port=19393, device="gpu")
 server.run_server()
